try to start protocol connection when IsAbleToEstablishProtocolConnectionWithUnity turns to true. refactor filesystemwatch in UnityEditorProtocol
